requirements.txt
fastapi
uvicorn
pydantic-ai
openai
python-multipart
python-dotenv
